requirements.txt
tensorflow
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
Pillow
deepface
mtcnn
